+++ pub-critic/package.json
@@ -16,6 +16,7 @@
     "eslint": "8.48.0",
     "eslint-config-next": "13.2.1",
     "next": "13.2.1",
+    "query-string": "^8.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.66.1",
