+++ pub-critic/src/app/api/GamesApi.ts
@@ -1,11 +1,17 @@
-import { apiKey, gamesApi } from "./GamesShared";
+import queryString from "query-string";
+export const gamesApi = "https://api.rawg.io/api";
+export const apiKey = "464bc085dbbf4f33bcb2ccb39d36a6ec";
 
 export const getGames = async (page: number = 1, pageSize: number = 20) => {
   try {
-    const response = await fetch(
-      `${gamesApi}/games?key=${apiKey}&page=${page}&page_size=${pageSize}`,
-      { next: { revalidate: 3600 * 24 * 31 } }
-    );
+    const query = queryString.stringify({
+      key: apiKey,
+      page,
+      page_size: pageSize,
+    });
+    const response = await fetch(`${gamesApi}/games?${query}`, {
+      next: { revalidate: 3600 * 24 * 31 },
+    });
     return response.json();
   } catch (error) {
     console.log(error);
@@ -17,10 +23,16 @@ export const getLatestGames = async (
   pageSize: number = 20
 ) => {
   try {
-    const response = await fetch(
-      `${gamesApi}/games?key=${apiKey}&dates=2023-01-01&page=${page}&page_size=${pageSize}&ordering=-released`,
-      { next: { revalidate: 3600 * 8 } }
-    );
+    const query = queryString.stringify({
+      key: apiKey,
+      dates: "2023-01-01",
+      page,
+      page_size: pageSize,
+      ordering: "-released",
+    });
+    const response = await fetch(`${gamesApi}/games?${query}`, {
+      next: { revalidate: 3600 * 8 },
+    });
     return response.json();
   } catch (error) {
     console.log(error);
@@ -32,10 +44,15 @@ export const getRatedGames = async (
   pageSize: number = 20
 ) => {
   try {
-    const response = await fetch(
-      `${gamesApi}/games?key=${apiKey}&page=${page}&page_size=${pageSize}&ordering=-metacritic`,
-      { next: { revalidate: 3600 * 8 } }
-    );
+    const query = queryString.stringify({
+      key: apiKey,
+      page,
+      page_size: pageSize,
+      ordering: "-metacritic",
+    });
+    const response = await fetch(`${gamesApi}/games?${query}`, {
+      next: { revalidate: 3600 * 8 },
+    });
     return response.json();
   } catch (error) {
     console.log(error);
