+++ pub-critic/src/app/components/GameCard/GameCard.module.scss
@@ -23,7 +23,7 @@
     }
     width: vw-calc(460);
     height: vw-calc(450);
-    border-radius: vw-calc(8) vw-calc(8) 0 0;
+    border-radius: vw-calc(8 8 0 0);
 
     > img {
       display: flex;
